* More cluster work :) * Changed logonserver packet size to use a size of 4 bytes instead of 2 bytes. This will solve the disconnection problem when sending character mappings on larger realms (damn, why didn't I think of that before)

+++ src/game/LogonCommClient.h
@@ -20,7 +20,7 @@
 
 class LogonCommClientSocket : public Socket
 {
-	uint16 remaining;
+	uint32 remaining;
 	uint16 opcode;
 	RC4Engine _sendCrypto;
 	RC4Engine _recvCrypto;
